requirements.txt
langchain_community
langchain_text_splitters
langchain.vectorstores
chromadb
sentence_transformers